Setup navigation, add basic navigation for game screen
Navigation is made to the ResultsScreen, changed the size of question in TriviaGameScreen, to do this added new TextStyle to Type.kt, in TriviaGameScreen add glow when user click an option, navigation to next question, and if there is no other questions go to ResultsScreen. Add dependencies for navigation, and update one dependency.